bulk edit metadata
