+++ pantheon-bundle/frontend/src/app/Utils.tsx
@@ -6,15 +6,15 @@ export class Utils {
      * @param options: object     
      */
 
-    public fetchHelper = (endpoint: string, options: object) => {
+    static fetchHelper = (endpoint: string, options: object) => {
     
         if (options !== null) {
             return fetch(endpoint, options)
-              .then(this.handleErrors)
+              .then(Utils.handleErrors)
               .then(response => response.json())
         } else {
             return fetch(endpoint)
-              .then(this.handleErrors)
+              .then(Utils.handleErrors)
               .then(response => response.json())
         }
     }
@@ -23,11 +23,32 @@ export class Utils {
      * handleErrors
      * @param response     
      */
-    public handleErrors = (response) => {
+    static handleErrors = (response) => {
         if (!response.ok) {
             throw Error(response.statusText)
         }
         return response
+    }
+
+    /**
+     * checks if a given api end point exists for a draft node.
+     * This method can be renamed to something more generic as it 
+     * can be used to check any nodepath
+     * @param path 
+     */
+    static draftExist(path) {
+        let exists = false
+        return fetch(path + ".json")
+            .then(response => {
+                if (response.ok) {
+                    exists = true
+                }
+                return exists
+            })
+            .catch((error) => {
+                console.log("[draftExist] error detected=>", error + " for " + path)
+                return false
+            })
     }
 }
