+++ pantheon-bundle/frontend/src/app/app.css
@@ -127,3 +127,9 @@ color: var(--pf-global--success-color--100)
     pointer-events: none;
     opacity: 0.4;
 }
+
+.p2-search__pf-c-alert{
+    position: absolute;
+    top:160px;
+    width: 50%;
+}
